Move database module to database folder
